Allow modules to specify database schema
This allows us to break down and shorten the master
data/database-schema.php file.

Note: that all pieces of the schema will be loaded even if a module is
disabled. This just allows for far less rebuilding of the database
even if modules are turned on or off.

This patch does not completely remove all the module pieces from the
database-schema.php but it makes a good first pass and establishes a
baseline for modules moving forward.