Block duplicated email
This transition is going to be a bit timing sensitive.

You will want to run the clean_duplicate_accounts tool via the web
when there is no neon update running.  Have ?update=1 to do the work
otherwise it will be a dry-run.

Then when the next neon update occures it should properly
not re-duplicate accounts and new neon accounts should not end up as
duplicates.

However if a neon update occures after this patch is applies and
before you run clean_duplicate_accounts, then the neon update will
behave a bit unpredictable around accounts with duplicate email and
may not assige the correct primary account.