Implement getMember to replace \lookup_xxx and findMember calls
This does an important thing with an important side effect:

    * Continues to disconnect us from the legacy PHP system

However also removes API level Neon support.

The removal of Neon support should be a non-issue because the legacy
web code should be importing all members into the database as they are
created.  But it is important to note that changes made to members via
the API will not propigate back into Neon either.